+++ src/common/dao/index.ts
@@ -1,2 +1,3 @@
 export { default as GovProposalCardHead } from "./GovProposalCardHead"
 export { default as DaoProposalsList } from "./DaoProposalsList"
+export { default as InsuranceAccidentMembersTable } from "./InsuranceAccidentMembersTable"
